Backend/personal-account/pyproject.toml
[tool.poetry]
name = "personal-account"
version = "0.1.0"
description = ""
authors = ["Nikita Lupenko <nicklupman@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.75.1"
uvicorn = "^0.17.6"
loguru = "^0.6.0"
psycopg2-binary = "^2.9.3"
pydantic = {extras = ["dotenv", "email"], version = "^1.9.0"}
SQLAlchemy = "^1.4.35"
alembic = "^1.7.7"
FastAPI-SQLAlchemy = "^0.2.1"
PyJWT = "^2.3.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.5"

[tool.poetry.dev-dependencies]
black = "^22.3.0"
flake8 = "^4.0.1"
mypy = "^0.942"
rope = "^1.0.0"
bandit = "^1.7.4"
pytest = "^7.1.1"
python-dateutil = "^2.8.2"
pytest-cov = "^3.0.0"
flake8-eradicate = "^1.2.0"
flake8-print = "^4.0.0"
flake8-use-fstring = "^1.3"
isort = "^5.10.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
